% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths.R
\docType{data}
\name{deaths}
\alias{deaths}
\title{Registered deaths in New Zealand by age, sex, and region, 1997-2016.}
\format{An array with dimensions "age", "sex", "region", and "time"}
\source{
Custom tabulation from Statistics New Zealand.
}
\usage{
deaths
}
\description{
The years are June years.  For instance, the year "1997" runs from 1 July 1996
to 30 Jun 1997.
}
\details{
To preserve confidentiality, death counts have been randomly rounded to base 3.
}
\keyword{datasets}
