% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.mig.R
\docType{data}
\name{external.mig}
\alias{external.mig}
\title{Permament and long-term arrivals to and departures from New Zealand,
1997-2016.}
\format{An array with dimensions "age", "sex", "region", and "time"}
\source{
Custom tabulation from Statistics New Zealand.
}
\usage{
external.mig
}
\description{
A permanent and long-term arrival is border crossing by a
person who intends to reside in the country for at least on year.
A permanent and long-term departure is a border crossing
by a person who intends to reside out of the country for at least
one year.
}
\details{
The years are June years.  For instance, the year "1997" runs from 1 July 1996
to 30 Jun 1997.

The data derive from international arrival and departure cards filled
out when entering and leaving the country.
}
\keyword{datasets}
