% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idi.erp.R
\docType{data}
\name{idi.erp}
\alias{idi.erp}
\title{Estimates of New Zealand population by age and sex from administrative
data, 2007-2016.}
\format{An array with dimensions "age", "sex", and "time"}
\source{
Statistics New Zealand website, downloaded 7 December 2017.
}
\usage{
idi.erp
}
\description{
The population counts refer to 30 June of each year. The data are
"Experimental population estimates from linked administrative data",
published in 2 October 2017. In the title "IDI-ERP", "IDI" means
"Integrated Data Infrastructure", and "ERP" means "Estimated
Resident Population".
}
\keyword{datasets}
