% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.mig.R
\docType{data}
\name{internal.mig}
\alias{internal.mig}
\title{Internal migration flows between census years, in New Zealand, 1996-2013.}
\format{An array with dimensions "age", "sex", "region_orig",
"region_dest", and "time"}
\source{
Custom tabulation from Statistics New Zealand.
}
\usage{
internal.mig
}
\description{
The data come from the "address 5 year ago" question in the population
census.  \code{"region_orig"} is the person's region 5 years before
the census, and \code{"region_dest"} is the person's region
at the time of the census.  The counts have been rounded to
multiples of 3.  Counts below 6 are suppressed in the original data,
to preserve confidentiality.  Values for these counts have been imputed,
by fitting a loglinear model to the data, and then randomly drawing
}
\keyword{datasets}
