% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{data}
\name{population}
\alias{population}
\title{Estimates of New Zealand population by age, sex, and region, 1996-2016.}
\format{An array with dimensions "age", "sex", "region", and "year"}
\source{
Custom tabulation from Statistics New Zealand.
}
\usage{
population
}
\description{
The population counts refer to 30 June of each year.
}
\keyword{datasets}
